Changed mileage to int

+++ src/no/hvl/dat109/controllers/ReturnController.java
@@ -34,7 +34,7 @@ public class ReturnController {
         LocalDateTime currentDate = LocalDateTime.now();
 
         System.out.println("Enter current mileage:");
-        double mileage = sc.nextDouble();
+        int mileage = sc.nextInt();
         sc.nextLine();
 
         Vehicle vehicle = reservation.getVehicle();
